profile/block-io: Generalized BlockIOParser
